Scripts: avoid restarting moonraker in githook
The policykit script restarts moonraker at the end, so if this
script is run as a moonraker post-merge git hook it will
interrupt the moonraker install process. The policykit script
now copies the script to tmp and comments out the systemctl
command and then returns a value if moonraker needs to be restarted
which is then used in the ratos-update script to restart moonraker
instead. This way moonraker updates will still work smoothly, in case
the user updated RatOS before updating Moonraker.