Scripts: make sure binaries are owned by pi
During image compilation the compile-binaries script
is run as root. Ownership of the firmware binary files are
now properly set in compile-binaries. This will fail
if files are not already owned by pi when running the macro
through Mainsail, just like before, but will be fixed for
future image releases.

Github: #41