M118 | Homing | ABL | MQTT
I added some features to get the printer faster to the point where it is ready to print, especially if you print multiple prints after an other. No Homing is needed if the printer was already homed after the last time the steppers were turned off f.e.

Same for Z_TILT_ADJUST.

M118 is shown at the top of Klipper Screen, M117 not

Nozzle Cleaning Option before the print starts

MQTT => for all guys out there that use MQTT and configured it we can send messages with my additions to the broker and do what ever we want with it. F.e. let Alexa tell us that our print is done or filament runout is detected... or just a simple pushmessage via. pushover or something like this.