Make configurable the end-print park's height
Even when parking at the rear, the EVA carriage is big enough that a
mere 2mm of Z separation makes the sheet removal process very precise
and precarious.

Since this will be subjective to different people's motor skills etc.;
instead of just changing the magic number, I propose factoring it out
like so.