Let the nozzle cool to 150 before probing the bed
The status quo means that the nozzle might be all the way at printing
temperature if two prints are run in quick succession (especially if
I have a second flexplate ready to go while the first one cools).

This is bad because temperature-dependent probes will get a weird result
as the probe will cool down during the probing; and because until the
nozzle cools enough, plastic will ooze onto the bed during the probing.

The solution, should it be desirable, is simple.