Factor most of the magic-numbers out of PRIME_LINE
This allows customization of the line's origin, direction and length.
Compared to the status quo, this should be a big help for the
macro-customizing type of person in case they mindlessly fill the slicer
plate with models and disregard the ~8mm of margin that the default
hardcoded origin (5,10) requires.

Additionally, this makes the line's E-values calculated entirely from
simple physical facts, including the nozzle bore which can be passed in
as a parameter if it's not 0.4mm, and the filament bore which is assumed
to be 1.75 (because Jinja2 does not import the math library so instead
I precomputed and hardcoded the filament's cross-sectional diameter).

Lastly, this includes a solution for an annoying problem where the
nozzle consistently goes a bit cold during the prime line,
causing SuperSlicer's insistently-inserted M109 to not immediately pass,
causing a long pause after the prime line which defeats the purpose of
the prime line by allowing the nozzle to deplete itself via oozing.

I've tested this on my V-Minion quite a bit, and it seems to produce
similar results as the status quo.