Update config.cfg to include z1 and z2 instead of extruder which is used from toolboard
When configuring a VCore3 or 4 the configuration is missing configuration for z1 and z2. In the configurator it states that when using this board a toolboard is required. Going through the steps and choosing the motors for z1 and z2 works fine, but at the last step it fails since there were no configurations defined for z1 and z2. 
This fixes that issue.